+++ sentry-rails/lib/sentry/rails/tracing/active_storage_subscriber.rb
@@ -19,9 +19,17 @@ module Sentry
           analyze.active_storage
         ].freeze
 
+        SPAN_ORIGIN = "auto.file.rails".freeze
+
         def self.subscribe!
           subscribe_to_event(EVENT_NAMES) do |event_name, duration, payload|
-            record_on_current_span(op: "file.#{event_name}".freeze, start_timestamp: payload[START_TIMESTAMP_NAME], description: payload[:service], duration: duration) do |span|
+            record_on_current_span(
+              op: "file.#{event_name}".freeze,
+              origin: SPAN_ORIGIN,
+              start_timestamp: payload[START_TIMESTAMP_NAME],
+              description: payload[:service],
+              duration: duration
+            ) do |span|
               payload.each do |key, value|
                 span.set_data(key, value) unless key == START_TIMESTAMP_NAME
               end