+++ sentry-rails/spec/sentry/rails/tracing/active_record_subscriber_spec.rb
@@ -34,6 +34,7 @@ RSpec.describe Sentry::Rails::Tracing::ActiveRecordSubscriber, :subscriber do
 
       span = transaction[:spans][0]
       expect(span[:op]).to eq("db.sql.active_record")
+      expect(span[:origin]).to eq("auto.db.rails")
       expect(span[:description]).to eq("SELECT \"posts\".* FROM \"posts\"")
       expect(span[:tags].key?(:cached)).to eq(false)
       expect(span[:trace_id]).to eq(transaction.dig(:contexts, :trace, :trace_id))
@@ -134,6 +135,7 @@ RSpec.describe Sentry::Rails::Tracing::ActiveRecordSubscriber, :subscriber do
 
       cached_query_span = transaction[:spans][1]
       expect(cached_query_span[:op]).to eq("db.sql.active_record")
+      expect(cached_query_span[:origin]).to eq("auto.db.rails")
       expect(cached_query_span[:description]).to eq("SELECT \"posts\".* FROM \"posts\"")
       expect(cached_query_span[:tags]).to include({ cached: true })