+++ sentry-resque/lib/sentry/resque.rb
@@ -15,6 +15,8 @@ module Sentry
     end
 
     class SentryReporter
+      SPAN_ORIGIN = "auto.queue.resque"
+
       class << self
         def record(queue, worker, payload, &block)
           Sentry.with_scope do |scope|
@@ -25,7 +27,13 @@ module Sentry
 
               name = contexts.dig(:"Active-Job", :job_class) || contexts.dig(:"Resque", :job_class)
               scope.set_transaction_name(name, source: :task)
-              transaction = Sentry.start_transaction(name: scope.transaction_name, source: scope.transaction_source, op: "queue.resque")
+              transaction = Sentry.start_transaction(
+                name: scope.transaction_name,
+                source: scope.transaction_source,
+                op: "queue.resque",
+                origin: SPAN_ORIGIN
+              )
+
               scope.set_span(transaction) if transaction
 
               yield